refactor(modules/events): dir contents restructure
