+++ src/modules/events/structures/handlers/index.ts
@@ -1,0 +1,1 @@
+export * from "./handler";
